requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
Pillow==9.4.0
numpy==1.24.2
Flask-Bootstrap==3.3.7.1
Bootstrap-Flask==2.2.0
Werkzeug==2.2.3